Use explicit integer division

+++ python/pencil/math/structure_function.py
@@ -8,8 +8,8 @@ import numpy as np
 import scipy as sp
 
 def structure_function(arr,y,x):
-    n=np.size(y)/2
-    m=np.size(x)/2
+    n=int(np.size(y)/2)
+    m=int(np.size(x)/2)
     D=np.zeros((n,m))
     for i in range(0, n, 1):
         yshift=np.roll(arr,i,axis=0)
@@ -17,7 +17,7 @@ def structure_function(arr,y,x):
             xshift=np.roll(yshift,j,axis=1)
             shift_diff=xshift-arr
             shift_squared=np.power(shift_diff,2)
-            shift_squared_cut=shift_squared[:,m/2:m]
+            shift_squared_cut=shift_squared[:,int(m/2):m]
             shift_squared_cut=\
                 shift_squared_cut[np.logical_not(np.isnan(shift_squared_cut))]
             if np.size(shift_squared_cut)==0:
@@ -28,7 +28,7 @@ def structure_function(arr,y,x):
 
 def structure_function_shift_range(arr,nmin,nmax,x):
     n=nmax-nmin
-    m=np.size(x)/2
+    m=int(np.size(x)/2)
     D=np.zeros((n,m))
     for i in range(nmin, nmax, 1):
         yshift=np.roll(arr,i,axis=0)
@@ -36,7 +36,7 @@ def structure_function_shift_range(arr,nmin,nmax,x):
             xshift=np.roll(yshift,j,axis=1)
             shift_diff=xshift-arr
             shift_squared=np.power(shift_diff,2)
-            shift_squared_cut=shift_squared[:,m/2:m]
+            shift_squared_cut=shift_squared[:,int(m/2):m]
             shift_squared_cut=\
                 shift_squared_cut[np.logical_not(np.isnan(shift_squared_cut))]
             if np.size(shift_squared_cut)==0:
